pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "control-panel"
version = "0.0.1"
description = ""
authors = [
    { name = "Dachboden", email = "info@dachboden.berlin" }
]
requires-python = ">=3.11"
dependencies = [
    "pygame-ce",
    "pyftdi",
    "moderngl",
    "pygame_shaders",
    "anaconsole @ git+https://github.com/AsinoEsel/anaconsole-pygame.git@main#egg=anaconsole",
    "artnet-python @ git+https://github.com/AsinoEsel/artnet-python.git@main#egg=artnet",
]

[project.optional-dependencies]
upy_dev = [
    "micropython-esp32-stubs==1.26.0.*",
    "pathspec",
]
server = [
    "RestrictedPython",
    "fastapi[standard]",
    "flake8 >= 5.0",
]
full = [
    "micropython-esp32-stubs==1.26.0.*",
    "pathspec",
    "RestrictedPython",
    "fastapi[standard]",
    "flake8 >= 5.0",
]


[project.entry-points."flake8.extension"]
RSP = "controlpanel.linting.restricted_checker:RestrictedPythonChecker"

[project.scripts]
controlpanel = "controlpanel.__main__:main"
generate_stubs = "dev_tools.generate_stubs.main:generate_all_stubs"
transfer = "dev_tools.transfer.transfer:main"
flash_firmware = "dev_tools.flash_firmware.flash_firmware:main"

[tool.setuptools.package-dir]
"" = "src"
userscripts = "userscripts"

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["controlpanel*", "dev_tools*", "userscripts*"]

[tool.setuptools.package-data]
"controlpanel.shared" = ["*.json"]
"controlpanel.game_manager" = ["icon.png"]
"controlpanel.shaders" = ["*.vert", "*.frag"]
"controlpanel.server" = [
    "templates/**/*.html",
    "static/**/*.*",
]